Домашняя работа №6

+++ src/main/java/com/bilalov/springeshoprb/dto/BucketDTO.java
@@ -13,8 +13,8 @@ import java.util.List;
 @AllArgsConstructor
 @Builder
 public class BucketDTO {
-    private int amountProducts;
-    private Double sum;
+    private long amountProducts;
+    private double sum;
     private List<BucketDetailDTO> bucketDetails = new ArrayList<>();
 
     public void aggregate() {
